# app/public/apidoc/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "NingBlog后台管理系统 API 接口文档",
    "version": "2.1.6",
    "title": "NingBlog 接口文档",
    "contact": {
      "email": "1700805832@qq.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "127.0.0.1",
  "basePath": "/api",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/content/getList": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "文档列表获取",
        "description": "根据合适条件查询文档列表",
        "operationId": "content_getList",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [{
            "name": "current",
            "in": "query",
            "description": "当前页码",
            "required": false,
            "type": "string"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页记录数",
            "required": false,
            "type": "string"
          },
          {
            "name": "searchkey",
            "in": "query",
            "description": "搜索关键字",
            "required": false,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "query",
            "description": "指定作者ID",
            "required": false,
            "type": "string"
          },
          {
            "name": "model",
            "in": "query",
            "description": "特殊分类 1:推荐帖子",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortby",
            "in": "query",
            "description": "按字段排序(0:默认按时间，1:热门)",
            "required": false,
            "type": "string",
            "default": "0"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Content"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        },
        "security": [{}]
      }
    },
...
    "/contentCategory/getList": {
      "get": {
        "tags": [
          "contentCategory"
        ],
        "summary": "获取文档类别列表",
        "description": "获取文档类别列表",
        "operationId": "contentCategory_getList",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ContentCategory"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        },
        "security": [{}]
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "Content": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "stitle": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "default": "1"
        },
        "categories": {
          "type": "array",
          "xml": {
            "name": "category",
            "wrapped": true
          },
          "items": {
            "$ref": "#/definitions/ContentCategory"
          }
        },
        "sortPath": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "xml": {
            "name": "contentTag",
            "wrapped": true
          },
          "items": {
            "$ref": "#/definitions/ContentTag"
          }
        },
        "keywords": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "sImg": {
          "type": "string",
          "default": "/upload/images/defaultImg.jpg"
        },
        "sImgType": {
          "type": "string",
          "default": "2"
        },
        "cover": {
          "type": "string"
        },
        "videoImg": {
          "type": "string",
          "default": ""
        },
        "discription": {
          "type": "string"
        },
        "appShowType": {
          "type": "string",
          "default": "1"
        },
        "imageArr": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "videoArr": {
          "type": "array",
          "xml": {
            "name": "videoArr",
            "wrapped": true
          },
          "items": {
            "type": "string"
          }
        },
        "duration": {
          "type": "string",
          "default": "0:01"
        },
        "date": {
          "type": "string"
        },
        "updateDate": {
          "type": "string"
        },
        "author": {
          "type": "string",
          "$ref": "#/definitions/AdminUser"
        },
        "uAuthor": {
          "type": "string",
          "$ref": "#/definitions/User"
        },
        "state": {
          "type": "string",
          "default": "0"
        },
        "draft": {
          "type": "string",
          "default": "0"
        },
        "dismissReason": {
          "type": "string"
        },
        "isTop": {
          "type": "number",
          "default": "0"
        },
        "roofPlacement": {
          "type": "string",
          "default": "0"
        },
        "clickNum": {
          "type": "number",
          "default": "1"
        },
        "comments": {
          "type": "string"
        },
        "simpleComments": {
          "type": "string"
        },
        "markDownComments": {
          "type": "string"
        },
        "commentNum": {
          "type": "number",
          "default": "0"
        },
        "likeNum": {
          "type": "number",
          "default": "0"
        }
      },
      "xml": {
        "name": "Content"
      }
    },
    "ContentCategory": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "uid": {
          "type": "number",
          "default": "0"
        },
        "name": {
          "type": "string"
        },
        "keywords": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "default": "1"
        },
        "sortId": {
          "type": "number",
          "default": "1"
        },
        "parentId": {
          "type": "string",
          "default": "0"
        },
        "enable": {
          "type": "boolean",
          "default": "true"
        },
        "date": {
          "type": "string"
        },
        "contentTemp": {
          "type": "string",
          "$ref": "#/definitions/TemplateItems"
        },
        "defaultUrl": {
          "type": "string",
          "default": ""
        },
        "homePage": {
          "type": "string",
          "default": "ui"
        },
        "sortPath": {
          "type": "string",
          "default": "0"
        },
        "comments": {
          "type": "string"
        },
        "sImg": {
          "type": "string"
        }
      },
      "xml": {
        "name": "ContentCategory"
      }
    },
    "ContentTag": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "alias": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "comments": {
          "type": "string"
        }
      },
      "xml": {
        "name": "ContentTag"
      }
    },
    "Message": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "contentId": {
          "type": "string",
          "$ref": "#/definitions/Content"
        },
        "contentTitle": {
          "type": "string"
        },
        "author": {
          "type": "string",
          "$ref": "#/definitions/User"
        },
        "adminAuthor": {
          "type": "string",
          "$ref": "#/definitions/Content"
        },
        "replyAuthor": {
          "type": "string",
          "$ref": "#/definitions/User"
        },
        "adminReplyAuthor": {
          "type": "string",
          "$ref": "#/definitions/AdminUser"
        },
        "state": {
          "type": "boolean",
          "default": false
        },
        "utype": {
          "type": "string",
          "default": "0"
        },
        "relationMsgId": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "praise_num": {
          "type": "number",
          "default": 0
        },
        "had_praise": {
          "type": "boolean",
          "default": false
        },
        "praiseMembers": {
          "type": "string"
        },
        "content": {
          "type": "string",
          "default": "输入评论内容..."
        }
      },
      "xml": {
        "name": "Message"
      }
    },
...
    "TemplateItems": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "forder": {
          "type": "string",
          "default": "2-stage-default"
        },
        "cateName": {
          "type": "string",
          "default": "contentList"
        },
        "detailName": {
          "type": "string",
          "default": "detail"
        },
        "isDefault": {
          "type": "boolean",
          "default": false
        },
        "date": {
          "type": "string"
        },
        "comment": {
          "type": "string"
        }
      },
      "xml": {
        "name": "TemplateItems"
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "enable": {
          "type": "boolean",
          "default": true
        },
        "password": {
          "type": "string"
        },
        "countryCode": {
          "type": "string"
        },
        "idType": {
          "type": "string",
          "default": "1"
        },
        "comments": {
          "type": "string"
        },
        "introduction": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "logo": {
          "type": "string",
          "default": "/static/upload/images/defaultlogo.png"
        },
        "group": {
          "type": "string",
          "default": "0"
        },
        "birth": {
          "type": "string",
          "default": "1770-01-01T00:00:00.000Z"
        },
        "gender": {
          "type": "string",
          "default": "0"
        },
        "despises": {
          "type": "array",
          "xml": {
            "name": "despises",
            "wrapped": true
          },
          "items": {
            "type": "string",
            "$ref": "#/definitions/Content"
          }
        },
        "despiseMessage": {
          "type": "array",
          "xml": {
            "name": "despiseMessage",
            "wrapped": true
          },
          "items": {
            "type": "string",
            "$ref": "#/definitions/Message"
          }
        },
        "favorites": {
          "type": "array",
          "xml": {
            "name": "favorites",
            "wrapped": true
          },
          "items": {
            "type": "string",
            "$ref": "#/definitions/Content"
          }
        },
        "praiseContents": {
          "type": "array",
          "xml": {
            "name": "praiseContents",
            "wrapped": true
          },
          "items": {
            "type": "string",
            "$ref": "#/definitions/Content"
          }
        },
        "praiseMessages": {
          "type": "array",
          "xml": {
            "name": "praiseMessages",
            "wrapped": true
          },
          "items": {
            "type": "string",
            "$ref": "#/definitions/Message"
          }
        },
        "state": {
          "default": "1"
        },
        "followers": {
          "type": "array",
          "xml": {
            "name": "followers",
            "wrapped": true
          },
          "items": {
            "type": "string",
            "$ref": "#/definitions/User"
          }
        },
        "watchers": {
          "type": "array",
          "xml": {
            "name": "followers",
            "wrapped": true
          },
          "items": {
            "type": "string",
            "$ref": "#/definitions/User"
          }
        },
        "watchTags": {
          "type": "array",
          "xml": {
            "name": "watchTags",
            "wrapped": true
          },
          "items": {
            "type": "string",
            "$ref": "#/definitions/ContentTag"
          }
        },
        "retrieve_time": {
          "type": "string"
        },
        "loginActive": {
          "type": "boolean",
          "default": false
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "AdminUser": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "countryCode": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "logo": {
          "type": "string",
          "default": "/static/upload/images/defaultlogo.png"
        },
        "enable": {
          "type": "boolean",
          "default": false
        },
        "state": {
          "type": "string",
          "default": "1"
        },
        "auth": {
          "type": "boolean",
          "default": false
        }
      },
      "xml": {
        "name": "ContentTag"
      }
    },
...
  },
...
}